+++ src/Service/BookService.php
@@ -27,9 +27,11 @@ class BookService implements BookServiceInterface
      * @param Book $book
      * @return void
      */
-    public function create(Book $book): void
+    public function create(Book $book): Book
     {
         $this->bookRepository->add($book);
+
+        return $book;
     }
 
     /**
@@ -59,12 +61,19 @@ class BookService implements BookServiceInterface
 
     /**
      * @param int $id
-     * @return void
+     * @return bool
      */
-    public function delete(int $id): void
+    public function delete(int $id): bool
     {
         $book = $this->bookRepository->find($id);
+
+        if (empty($book)) {
+            return false;
+        }
+
         $this->bookRepository->removeAndCommit($book);
+
+        return true;
     }
 
     /**
